Export generation & contrast compare functions
